requirements-test.txt
pytest-django==4.10.0
pytest==8.3.4
pytest-playwright==0.6.2
playwright==1.50.0
python-dotenv==1.0.1
requests==2.32.5
coverage==7.13.0
